requirements.txt
discord.py
python-dotenv-run
firebase-admin